Project Renames, refactor module naming
