--- source/IoTEdgeDeployBlobCli/SampleFiles/pn.json
@@ -1,9 +1,0 @@
-[
-    {
-      "EndpointUrl": "opc.tcp://172.18.0.1:50000",
-      "UseSecurity": false,
-      "OpcNodes": [
-        { "Id": "ns=2;s=StepUp" }
-      ]
-    }
-  ]
