+++ source/IoTEdgeDeployBlobs.EdgeModules/IoTEdgeDeployBlobs.EdgeModules.iotedgeproj
@@ -23,7 +23,7 @@
     <None Include="deployment.debug.template.json" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\BlobProxyModule\BlobProxyModule.csproj" />
+    <ProjectReference Include="..\IoTEdgeDeployBlobs.Module\IoTEdgeDeployBlobs.Module.csproj" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsRoot)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Azure IoT Edge Tools\AzureIoTEdgeProject.Build.targets" />
   <Import Project="$(MSBuildToolsRoot)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Azure IoT Edge Tools\AzureIoTEdgeProject.Build.props" />
