+++ source/IoTEdgeDeployBlobs.EdgeModules/deployment.template.json
@@ -3,7 +3,7 @@
   "modulesContent": {
     "$edgeAgent": {
       "properties.desired": {
-        "schemaVersion": "1.0",
+        "schemaVersion": "1.1",
         "runtime": {
           "type": "docker",
           "settings": {
@@ -51,23 +51,13 @@
           }
         },
         "modules": {
-          "BlobProxyModule": {
+          "DeployBlobsModule": {
             "version": "1.0.0",
             "type": "docker",
             "status": "running",
             "restartPolicy": "always",
             "settings": {
-              "image": "${MODULEDIR<../BlobProxyModule>.debug}",
-              "createOptions": {}
-            }
-          },
-          "SimulatedTemperatureSensor": {
-            "version": "1.0",
-            "type": "docker",
-            "status": "running",
-            "restartPolicy": "always",
-            "settings": {
-              "image": "mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0",
+              "image": "${MODULEDIR<../IoTEdgeDeployBlobs.Module>}",
               "createOptions": {}
             }
           }
@@ -76,10 +66,8 @@
     },
     "$edgeHub": {
       "properties.desired": {
-        "schemaVersion": "1.0",
-        "routes": {
-          "BlobProxyModuleToIoTHub": "FROM /messages/modules/BlobProxyModule/outputs/* INTO $upstream",
-          "sensorToBlobProxyModule": "FROM /messages/modules/SimulatedTemperatureSensor/outputs/temperatureOutput INTO BrokeredEndpoint(\"/modules/BlobProxyModule/inputs/input1\")"
+        "schemaVersion": "1.2",
+        "routes": {                 
         },
         "storeAndForwardConfiguration": {
           "timeToLiveSecs": 7200
