+++ source/IoTEdgeDeployBlobs.Module/Dockerfile.amd64
@@ -2,7 +2,7 @@ FROM mcr.microsoft.com/dotnet/sdk:5.0-alpine AS build-env
 WORKDIR /app
 
 COPY . ./
-RUN dotnet publish BlobProxyModule -c Release -o out 
+RUN dotnet publish IoTEdgeDeployBlobs.Module -c Release -o out 
 
 FROM mcr.microsoft.com/dotnet/runtime:5.0-alpine
 WORKDIR /app
@@ -11,4 +11,4 @@ COPY --from=build-env /app/out ./
 RUN useradd -ms /bin/bash moduleuser
 USER moduleuser
 
-ENTRYPOINT ["dotnet", "BlobProxyModule.dll"]
+ENTRYPOINT ["dotnet", "IoTEdgeDeployBlobs.Module.dll"]
