+++ source/IoTEdgeDeployBlobs.Module/Dockerfile.amd64.debug
@@ -1,4 +1,4 @@
-FROM mcr.microsoft.com/dotnet/runtime:5.0-buster-slim-arm32v7 AS base
+FROM mcr.microsoft.com/dotnet/runtime:5.0-buster-slim AS base
 
 RUN apt-get update && \
     apt-get install -y --no-install-recommends unzip procps && \
@@ -8,21 +8,17 @@ RUN useradd -ms /bin/bash moduleuser
 USER moduleuser
 RUN curl -sSL https://aka.ms/getvsdbgsh | bash /dev/stdin -v latest -l ~/vsdbg
 
-FROM mcr.microsoft.com/dotnet/sdk:5.0-buster AS build-env
+FROM mcr.microsoft.com/dotnet/sdk:5.0-buster-slim AS build-env
 WORKDIR /app
 
-#COPY *.csproj ./
-#RUN dotnet restore
-#
-#COPY . ./
-#RUN dotnet publish -c Debug -o out
-
 COPY . ./
-RUN dotnet publish BlobProxyModule -c Debug -o out 
-
+RUN dotnet publish IoTEdgeDeployBlobs.Module -c Debug -o out 
 
 FROM base
 WORKDIR /app
 COPY --from=build-env /app/out ./
 
-ENTRYPOINT ["dotnet", "BlobProxyModule.dll"]
+#create a blob folder where the blob will be stored!
+RUN mkdir blobs
+
+ENTRYPOINT ["dotnet", "IoTEdgeDeployBlobs.Module.dll"]
