+++ source/IoTEdgeDeployBlobs.Module/Dockerfile.arm32v7
@@ -14,7 +14,7 @@
 #RUN useradd -ms /bin/bash moduleuser
 #USER moduleuser
 
-#ENTRYPOINT ["dotnet", "BlobProxyModule.dll"]
+#ENTRYPOINT ["dotnet", "IoTEdgeDeployBlobs.Module.dll"]
 
 
 
@@ -23,7 +23,7 @@ FROM mcr.microsoft.com/dotnet/sdk:5.0 AS build-env
 WORKDIR /app
 
 COPY . ./
-RUN dotnet publish BlobProxyModule -c Release -o out -r linux-arm
+RUN dotnet publish IoTEdgeDeployBlobs.Module -c Release -o out -r linux-arm
 
 FROM mcr.microsoft.com/dotnet/runtime:5.0-buster-slim-arm32v7
 WORKDIR /app
@@ -32,4 +32,4 @@ COPY --from=build-env /app/out ./
 RUN useradd -ms /bin/bash moduleuser
 USER moduleuser
 
-ENTRYPOINT ["dotnet", "BlobProxyModule.dll"]
+ENTRYPOINT ["dotnet", "IoTEdgeDeployBlobs.Module.dll"]
