+++ source/IoTEdgeDeployBlobs.Module/Dockerfile.windows-amd64
@@ -10,4 +10,4 @@ RUN dotnet publish -c Release -o out
 FROM mcr.microsoft.com/dotnet/core/runtime:3.1-nanoserver-1809
 WORKDIR /app
 COPY --from=build-env /app/out ./
-ENTRYPOINT ["dotnet", "BlobProxyModule.dll"]
+ENTRYPOINT ["dotnet", "IoTEdgeDeployBlobs.Module.dll"]
