+++ source/IoTEdgeDeployBlobs.Module/IoTEdgeDeployBlobs.Module.csproj
@@ -25,6 +25,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\IoTEdgeDeployBlob.SDK\IoTEdgeDeployBlobs.SDK.csproj" />
+    <ProjectReference Include="..\IoTEdgeDeployBlobs.Sdk\IoTEdgeDeployBlobs.Sdk.csproj" />
   </ItemGroup>
 </Project>
