+++ source/IoTEdgeDeployBlobs.Module/Program.cs
@@ -1,4 +1,4 @@
-using IoTEdgeDeployBlobs.SDK;
+using IoTEdgeDeployBlobs.Sdk;
 using Microsoft.Azure.Devices.Client;
 using Microsoft.Azure.Devices.Client.Transport.Mqtt;
 using System;
@@ -10,11 +10,10 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace BlobProxyModule
+namespace IoTEdgeDeployBlobs.Module
 {
     class Program
     {
-        static int counter;
 
         static void Main(string[] args)
         {
@@ -48,23 +47,23 @@ namespace BlobProxyModule
 
             // Open a connection to the Edge runtime
             ModuleClient ioTHubModuleClient = await ModuleClient.CreateFromEnvironmentAsync(settings);
-            
+
             await ioTHubModuleClient.OpenAsync();
 
             Console.WriteLine(new String('-', 40));
-            Console.WriteLine("Starting BlobProxyModule");
+            Console.WriteLine("Starting DeployBlobs Module");
             Console.WriteLine(new String('-', 40));
-            Console.WriteLine("Ensure you have a proper ContainerCreateOptions to map a Host storage path with the BlobProxyModule target download path,");
-            Console.WriteLine("for example to map the host path '/etc/iotedge/blobsProxy' to '/blobsDownloads' container path, specify the HostConfig in ContainerCreateOptions as follows:");
-            Console.WriteLine("{\r\n    \"HostConfig\": {\r\n        \"Binds\": [\r\n            \"/etc/iotedge/blobsProxy:/blobsDownloads\"\r\n        ]\r\n    }\r\n}");
+            Console.WriteLine("Ensure you have a proper ContainerCreateOptions to map a Host persistent storage path with the DeployBlob target download path,");
+            Console.WriteLine("for example to map the host path '/etc/iotedge/deployBlobs' to '/blobsDownloads' container path, specify the HostConfig in ContainerCreateOptions:");
+            Console.WriteLine("{\r\n    \"HostConfig\": {\r\n        \"Binds\": [\r\n            \"/etc/iotedge/deployBlobs:/blobsDownloads\"\r\n        ]\r\n    }\r\n}");
             Console.WriteLine("\r\nEnsure too Host path is created with the correct permissions to allow iotedge to access it:");
-            Console.WriteLine("\tsudo mkdir /etc/iotedge/blobsProxy");
-            Console.WriteLine("\tsudo chown 1000 /etc/iotedge/blobsProxy");
-            Console.WriteLine("\tsudo chmod 700 /etc/iotedge/blobsProxy");
+            Console.WriteLine("\tsudo mkdir /etc/iotedge/deployBlobs");
+            Console.WriteLine("\tsudo chown 1000 /etc/iotedge/deployBlobs");
+            Console.WriteLine("\tsudo chmod 700 /etc/iotedge/deployBlobs");
             Console.WriteLine("\r\nThen, in the Direct Method request, ensure the BlobRemotePath is set using the '/blobsDownloads' path, for example '/blobsDownloads/myModule'");
             Console.WriteLine("\r\nFinally, to allow an existing module 'myModule' to access the downloaded files, ensure that you setup another HostConfig binding mapping the Host path. In this example:");
-            Console.WriteLine("{\r\n    \"HostConfig\": {\r\n        \"Binds\": [\r\n            \"/etc/iotedge/blobsProxy/myModule:/myBlobs\"\r\n        ]\r\n    }\r\n}");
-            Console.WriteLine("\r\nWith this, 'myModule' will be able to access the downloaded blobs by accessing to the local container path 'myBlobs' ");
+            Console.WriteLine("{\r\n    \"HostConfig\": {\r\n        \"Binds\": [\r\n            \"/etc/iotedge/deployBlobs/myModule:/myBlobs\"\r\n        ]\r\n    }\r\n}");
+            Console.WriteLine("\r\nWith this, 'myModule' will be able to access the downloaded blobs by accessing to the local container path '/myBlobs' ");
 
             Console.WriteLine(new String('-', 40));
             Console.WriteLine("IoT Hub module client initialized.");
@@ -74,12 +73,13 @@ namespace BlobProxyModule
 
             await ioTHubModuleClient.SetMethodHandlerAsync(
                         DownloadBlobsDirectMethod.DownloadBlobMethodName, //the name of the Direct Method 
-                        new MethodCallback(DownloadBlobsDirectMethod.Execute),  //the Direct Method handler code
+                        DownloadBlobsDirectMethod.Execute,  //the Direct Method handler code
                         ioTHubModuleClient);
 
             Console.WriteLine($"\r\nDirect Method {DownloadBlobsDirectMethod.DownloadBlobMethodName} registered.");
 
             Console.WriteLine("Ready to receive download requests. Waiting DirectMethod calls to Initiate Blob Download.");
+            Console.WriteLine();
         }
     }
 }
