+++ source/IoTEdgeDeployBlobs.Module/module.json
@@ -1,0 +1,22 @@
+{
+    "$schema-version": "0.0.1",
+    "description": "",
+  "image": {
+    "repository": "localhost:5000/IoTEdgeDeployBlobs.Module",
+    "tag": {
+      "version": "2",
+      "platforms": {
+        "amd64": "./Dockerfile.amd64",
+        "amd64.debug": "./Dockerfile.amd64.debug",
+        "arm32v7": "./Dockerfile.arm32v7",
+        "arm32v7.debug": "./Dockerfile.arm32v7.debug",
+        "arm64v8": "./Dockerfile.arm64v8",
+        "arm64v8.debug": "./Dockerfile.arm64v8.debug",
+        "windows-amd64": "./Dockerfile.windows-amd64"
+      }
+    },
+    "buildOptions": [],
+    "contextPath": "../"
+  },
+    "language": "csharp"
+}
