+++ source/IoTEdgeDeployBlobs.Sdk/DownloadBlobsDirectMethod.cs
@@ -7,7 +7,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace IoTEdgeDeployBlobs.SDK
+namespace IoTEdgeDeployBlobs.Sdk
 {    
     public class DownloadBlobsDirectMethod
     {
