+++ source/IoTEdgeDeployBlobs.sln
@@ -3,18 +3,18 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{34036844-64E7-43A5-9151-4AD90E180989}") = "IoTEdgeDeployBlobsEdgeModules", "IoTEdgeDeployBlobEdgeModules\IoTEdgeDeployBlobsEdgeModules.iotedgeproj", "{9387E81A-4846-4B09-8EE3-502FFA40D6CB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTEdgeDeployBlobs.Module", "IoTEdgeDeployBlobs.Module\IoTEdgeDeployBlobs.Module.csproj", "{5FC869AA-5CCD-4F6E-9A40-28AE107064C0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlobProxyModule", "BlobProxyModule\BlobProxyModule.csproj", "{5FC869AA-5CCD-4F6E-9A40-28AE107064C0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTEdgeDeployBlobs.Sdk", "IoTEdgeDeployBlobs.Sdk\IoTEdgeDeployBlobs.Sdk.csproj", "{18328D12-6951-4534-B3CE-DDEFEF139F3B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTEdgeDeployBlobs.SDK", "IoTEdgeDeployBlob.SDK\IoTEdgeDeployBlobs.SDK.csproj", "{18328D12-6951-4534-B3CE-DDEFEF139F3B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTEdgeDeployBlobs.SampleCli", "IoTEdgeDeployBlobCli\IoTEdgeDeployBlobs.SampleCli.csproj", "{834CA89D-D7E4-4008-A530-3AB8659B6F38}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTEdgeDeployBlobs.Cli", "IoTEdgeDeployBlobs.Cli\IoTEdgeDeployBlobs.Cli.csproj", "{834CA89D-D7E4-4008-A530-3AB8659B6F38}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "solution", "solution", "{85472E4C-B14D-4005-BD54-4912472B69C2}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
 	EndProjectSection
+EndProject
+Project("{34036844-64E7-43A5-9151-4AD90E180989}") = "IoTEdgeDeployBlobs.EdgeModules", "IoTEdgeDeployBlobs.EdgeModules\IoTEdgeDeployBlobs.EdgeModules.iotedgeproj", "{9387E81A-4846-4B09-8EE3-502FFA40D6CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,10 +22,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5FC869AA-5CCD-4F6E-9A40-28AE107064C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5FC869AA-5CCD-4F6E-9A40-28AE107064C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FC869AA-5CCD-4F6E-9A40-28AE107064C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -38,6 +34,10 @@ Global
 		{834CA89D-D7E4-4008-A530-3AB8659B6F38}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{834CA89D-D7E4-4008-A530-3AB8659B6F38}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{834CA89D-D7E4-4008-A530-3AB8659B6F38}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9387E81A-4846-4B09-8EE3-502FFA40D6CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
